remove Archive from repo
This folder is only used temporary when I need to apply major change to a file and I want to start from scratch while having a copy of the original file for reference (some pieces of code are difficult or long to rewrite). No need of this folder in the repo.