stop monitoring when BaM gets stuck
This is an unknown issue in BaM calibration. On random calibration run, BaM gets stuck with no clue regarding the reason. If this is not detected, moniroting continues endlessly. The code added in this commit does the following: it checks the calibration progress and if the same percentage value is obtained more than 10 times, BaM is considered stuck and monitoring stops with an error message inviting the user to "try again".